feat!: enhance annotator CLI options (#449)
* Make the input VCF path an
[argument](https://click.palletsprojects.com/en/8.1.x/arguments/) rather
than an [option](https://click.palletsprojects.com/en/8.1.x/options/). I
think this better matches the intent of the tool (and it was the only
required option, so might as well make it an arg, right?). Left it at
`nargs=1` but I think there's a case for making it `nargs=-1`... but
going 1 -> -1 wouldn't be a breaking change imo so I think it's fine to
put that decision off.
* pathlike args should be `click.Path`/`pathlib.Path`. Add additional
configurations, like the input file must exist and the output files must
be writable. Didn't make changes further up the stack because I wanted
to confine this PR to just the Click interface.
* remove some superfluous/inactive configs
* Touch up help text