feat!: show more informative error msg for vcf annotate output args issue (#520)
Minor issue. Previously, if you forgot to add an output option, the
error message you get concerns arg names passed to the `annotate()`
method, not the CLI itself, so not very helpful.

I also think it might make sense to affix default behavior to the CLI
(maybe just make a vcf in the current working dir) rather than raising
this error, but that's a slightly bigger change.