feat: define VRS schema version in package (#543)
supercedes #467. In that PR, I defined a hook for setuptools that would
use the tag info in a submodule and auto-update a variable with the
latest VRS schema tag, so that wheels built by our release workflow
would contain the right value. There were probably going to be some
hiccups that way, though, and I also couldn't get it to be testable in
github actions on account of how actions/checkout seemed to be pulling
the git repos.

Here, we just define a little old constant with the schema version. This
is a fully manual process, but theoretically we won't need to update it
that often, so it should be fine.