Integrate Hackathon vrs-in-a-box into vrs-python (#555)
Moved code from the Hackathon vrs-in-a-box project over in to a
subdirectory within vrs-python.

The code includes full instructions for how to create a vrs-in-a-box
Docker image from scratch
and examples for how to run using Docker.

The following items are (probably) still needed to bring this work to
completion:
- Create Docker images, push them to the ga4gh Docker hub and update the
documentation with image locations
- Test that vrs-annotate options besides --vcf-out and --ndjson-out work
as expected
- Test the sample WDL to make sure it works and maybe add additional
parameters supported by vrs-annotate